Social_networking/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==5.0
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
PyJWT==2.8.0
python-memcached==1.59
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
